refactor(home): split home_page into per-widget files
Promotes the previously inline _HomeAppBar / _HomeBottomBar / _HomeBody
/ _SettingsDrawer private classes to top-level widgets in
pages/home/widgets/. Page file shrinks from ~310 lines to ~60 lines of
pure state + composition.

home/widgets/:
- home_app_bar.dart       HomeAppBar (with the FittedBox calendar fix)
- home_bottom_bar.dart    HomeBottomBar
- home_body.dart          HomeBody (uses shared BlurredBackground)
- settings_drawer.dart    SettingsDrawer (uses shared
                          SettingSwitchRow + AppDrawerButton)

app.dart and test/.../home_page_test.dart pick up the new import path.
